+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- next-header -->
 
 ## [Unreleased] - ReleaseDate
+### Fixed
+- cursor: handle errors sent at the end of a response ([#56]).
+
+[#56]: https://github.com/loyd/clickhouse.rs/issues/56
 
 ## [0.11.2] - 2023-01-03
 ### Added